Move descriptor_set.rs to parent, make descriptor.rs its child
