+++ vulkano/src/command_buffer/auto.rs
@@ -37,8 +37,8 @@ use crate::command_buffer::SecondaryCommandBuffer;
 use crate::command_buffer::StateCacher;
 use crate::command_buffer::StateCacherOutcome;
 use crate::command_buffer::SubpassContents;
-use crate::descriptor::descriptor::{DescriptorBufferDesc, DescriptorDescTy};
-use crate::descriptor::descriptor_set::{DescriptorSetDesc, DescriptorSetsCollection};
+use crate::descriptor_set::descriptor::{DescriptorBufferDesc, DescriptorDescTy};
+use crate::descriptor_set::{DescriptorSetDesc, DescriptorSetsCollection};
 use crate::device::Device;
 use crate::device::DeviceOwned;
 use crate::device::Queue;
