+++ vulkano/src/command_buffer/synced/commands.rs
@@ -25,9 +25,9 @@ use crate::command_buffer::CommandBufferExecError;
 use crate::command_buffer::ImageUninitializedSafe;
 use crate::command_buffer::SecondaryCommandBuffer;
 use crate::command_buffer::SubpassContents;
-use crate::descriptor::descriptor::DescriptorDescTy;
-use crate::descriptor::descriptor::ShaderStages;
-use crate::descriptor::descriptor_set::DescriptorSet;
+use crate::descriptor_set::descriptor::DescriptorDescTy;
+use crate::descriptor_set::descriptor::ShaderStages;
+use crate::descriptor_set::DescriptorSet;
 use crate::format::ClearValue;
 use crate::image::ImageAccess;
 use crate::image::ImageLayout;
