+++ vulkano/src/command_buffer/sys.rs
@@ -17,8 +17,8 @@ use crate::command_buffer::CommandBufferLevel;
 use crate::command_buffer::CommandBufferUsage;
 use crate::command_buffer::SecondaryCommandBuffer;
 use crate::command_buffer::SubpassContents;
-use crate::descriptor::descriptor::ShaderStages;
-use crate::descriptor::descriptor_set::UnsafeDescriptorSet;
+use crate::descriptor_set::descriptor::ShaderStages;
+use crate::descriptor_set::UnsafeDescriptorSet;
 use crate::device::Device;
 use crate::device::DeviceOwned;
 use crate::format::ClearValue;
