+++ vulkano/src/descriptor_set/persistent.rs
@@ -9,20 +9,20 @@
 
 use crate::buffer::BufferAccess;
 use crate::buffer::BufferViewRef;
-use crate::descriptor::descriptor::DescriptorDesc;
-use crate::descriptor::descriptor::DescriptorDescTy;
-use crate::descriptor::descriptor::DescriptorImageDesc;
-use crate::descriptor::descriptor::DescriptorImageDescArray;
-use crate::descriptor::descriptor::DescriptorImageDescDimensions;
-use crate::descriptor::descriptor::DescriptorType;
-use crate::descriptor::descriptor_set::DescriptorPool;
-use crate::descriptor::descriptor_set::DescriptorPoolAlloc;
-use crate::descriptor::descriptor_set::DescriptorSet;
-use crate::descriptor::descriptor_set::DescriptorSetDesc;
-use crate::descriptor::descriptor_set::DescriptorWrite;
-use crate::descriptor::descriptor_set::StdDescriptorPoolAlloc;
-use crate::descriptor::descriptor_set::UnsafeDescriptorSet;
-use crate::descriptor::descriptor_set::UnsafeDescriptorSetLayout;
+use crate::descriptor_set::descriptor::DescriptorDesc;
+use crate::descriptor_set::descriptor::DescriptorDescTy;
+use crate::descriptor_set::descriptor::DescriptorImageDesc;
+use crate::descriptor_set::descriptor::DescriptorImageDescArray;
+use crate::descriptor_set::descriptor::DescriptorImageDescDimensions;
+use crate::descriptor_set::descriptor::DescriptorType;
+use crate::descriptor_set::DescriptorPool;
+use crate::descriptor_set::DescriptorPoolAlloc;
+use crate::descriptor_set::DescriptorSet;
+use crate::descriptor_set::DescriptorSetDesc;
+use crate::descriptor_set::DescriptorWrite;
+use crate::descriptor_set::StdDescriptorPoolAlloc;
+use crate::descriptor_set::UnsafeDescriptorSet;
+use crate::descriptor_set::UnsafeDescriptorSetLayout;
 use crate::device::Device;
 use crate::device::DeviceOwned;
 use crate::format::Format;
