+++ vulkano/src/lib.rs
@@ -77,7 +77,7 @@ mod tests;
 mod extensions;
 pub mod buffer;
 pub mod command_buffer;
-pub mod descriptor;
+pub mod descriptor_set;
 pub mod device;
 pub mod format;
 mod version;
