+++ vulkano/src/pipeline/layout/desc.rs
@@ -7,12 +7,12 @@
 // notice may not be copied, modified, or distributed except
 // according to those terms.
 
-use crate::descriptor::descriptor::DescriptorBufferDesc;
-use crate::descriptor::descriptor::DescriptorDesc;
-use crate::descriptor::descriptor::DescriptorDescSupersetError;
-use crate::descriptor::descriptor::DescriptorDescTy;
-use crate::descriptor::descriptor::ShaderStages;
-use crate::descriptor::descriptor_set::DescriptorSetsCollection;
+use crate::descriptor_set::descriptor::DescriptorBufferDesc;
+use crate::descriptor_set::descriptor::DescriptorDesc;
+use crate::descriptor_set::descriptor::DescriptorDescSupersetError;
+use crate::descriptor_set::descriptor::DescriptorDescTy;
+use crate::descriptor_set::descriptor::ShaderStages;
+use crate::descriptor_set::DescriptorSetsCollection;
 use crate::device::Device;
 use crate::pipeline::layout::limits_check;
 use fnv::FnvHashSet;
@@ -418,7 +418,7 @@ impl fmt::Display for PipelineLayoutNotSupersetError {
 
 #[cfg(test)]
 mod tests {
-    use crate::descriptor::descriptor::ShaderStages;
+    use crate::descriptor_set::descriptor::ShaderStages;
     use crate::pipeline::layout::PipelineLayoutDesc;
     use crate::pipeline::layout::PipelineLayoutDescError;
     use crate::pipeline::layout::PipelineLayoutDescPcRange;
