+++ vulkano/src/pipeline/layout/limits_check.rs
@@ -9,8 +9,8 @@
 
 //! Contains the `check_desc_against_limits` function and the `PipelineLayoutLimitsError` error.
 
-use crate::descriptor::descriptor::DescriptorType;
-use crate::descriptor::descriptor::ShaderStages;
+use crate::descriptor_set::descriptor::DescriptorType;
+use crate::descriptor_set::descriptor::ShaderStages;
 use crate::device::Properties;
 use crate::pipeline::layout::PipelineLayoutDesc;
 use crate::pipeline::layout::PipelineLayoutDescPcRange;
